requirements.txt
absl-py==1.0.0
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cmake==3.22.4
colorama==0.4.4
comtypes==1.1.11
cycler==0.11.0
fonttools==4.33.3
gTTS==2.2.4
idna==3.3
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.2
mediapipe==0.8.10
numpy==1.22.4
ohbot==4.0.8
ohbotWin==1.14
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pip==21.3.1
playsound==1.3.0
protobuf==3.20.1
pydub==0.25.1
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
requests==2.27.1
setuptools==60.2.0
six==1.16.0
urllib3==1.26.9
wheel==0.37.1
eel~=0.15.0
torch~=1.13.0
torchvision~=0.14.0
pandas~=1.5.2
soundfile~=0.11.0